Create demo5 for payments thru channel
Add another account in accounts.yml, this
account is the channel account, which is
used for the tx.

Remember to edit the stellar-sdk gem bec-
ause as of now (v0.5.0), the operation
src address is the same as tx src add.
More details can be found at the
demo5.rb.